[watchos][spritekit] Enable SpriteKit for watchOS
A few things are left partially disabled as they requires SceneKit and AVFoundation

!missing-selector! SK3DNode::autoenablesDefaultLighting not bound
!missing-selector! SK3DNode::hitTest:options: not bound
!missing-selector! SK3DNode::isPlaying not bound
!missing-selector! SK3DNode::loops not bound
!missing-selector! SK3DNode::pointOfView not bound
!missing-selector! SK3DNode::sceneTime not bound
!missing-selector! SK3DNode::scnScene not bound
!missing-selector! SK3DNode::setAutoenablesDefaultLighting: not bound
!missing-selector! SK3DNode::setLoops: not bound
!missing-selector! SK3DNode::setPlaying: not bound
!missing-selector! SK3DNode::setPointOfView: not bound
!missing-selector! SK3DNode::setSceneTime: not bound
!missing-selector! SK3DNode::setScnScene: not bound
!missing-selector! SKAudioNode::avAudioNode not bound
!missing-selector! SKAudioNode::initWithAVAudioNode: not bound
!missing-selector! SKAudioNode::initWithFileNamed: not bound
!missing-selector! SKAudioNode::initWithURL: not bound
!missing-selector! SKAudioNode::setAvAudioNode: not bound
!missing-selector! SKScene::audioEngine not bound
